backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.10.1
PyJWT==2.10.1
python-multipart==0.0.20
email-validator==2.3.0
psycopg[binary]==3.2.9
pypdf==5.9.0
python-docx==1.2.0
openai==1.108.1
